Move discord to server folder
